refactor unit test into group and object directories
