+++ test/unit/objects/lke_test.py
@@ -7,7 +7,6 @@ from linode_api4.objects import (
     LKECluster,
     LKEClusterControlPlaneACLAddressesOptions,
     LKEClusterControlPlaneACLOptions,
-    LKEClusterControlPlaneOptions,
     LKENodePool,
 )
 from linode_api4.objects.lke import LKENodePoolNode, LKENodePoolTaint
@@ -165,36 +164,6 @@ class LKETest(ClientBaseCase):
         self.assertIsNotNone(pool.nodes)
         self.assertIsNotNone(pool.autoscaler)
         self.assertIsNotNone(pool.tags)
-
-    def test_cluster_create_with_acl(self):
-        """
-        Tests that an LKE cluster can be created with a control plane ACL configuration.
-        """
-
-        with self.mock_post("lke/clusters") as m:
-            self.client.lke.cluster_create(
-                "us-mia",
-                "test-acl-cluster",
-                [self.client.lke.node_pool("g6-nanode-1", 3)],
-                "1.29",
-                control_plane=LKEClusterControlPlaneOptions(
-                    acl=LKEClusterControlPlaneACLOptions(
-                        enabled=True,
-                        addresses=LKEClusterControlPlaneACLAddressesOptions(
-                            ipv4=["10.0.0.1/32"], ipv6=["1234::5678"]
-                        ),
-                    )
-                ),
-            )
-
-            assert "high_availability" not in m.call_data["control_plane"]
-            assert m.call_data["control_plane"]["acl"]["enabled"]
-            assert m.call_data["control_plane"]["acl"]["addresses"]["ipv4"] == [
-                "10.0.0.1/32"
-            ]
-            assert m.call_data["control_plane"]["acl"]["addresses"]["ipv6"] == [
-                "1234::5678"
-            ]
 
     def test_cluster_get_acl(self):
         """
