+++ test/unit/objects/placement_test.py
@@ -1,11 +1,9 @@
 from test.unit.base import ClientBaseCase
 
-from linode_api4 import PlacementGroupPolicy
 from linode_api4.objects import (
     MigratedInstance,
     PlacementGroup,
     PlacementGroupMember,
-    PlacementGroupType,
 )
 
 
@@ -24,46 +22,6 @@ class PlacementTest(ClientBaseCase):
 
         self.validate_pg_123(pg)
         assert pg._populated
-
-    def test_list_pgs(self):
-        """
-        Tests that you can list PGs.
-        """
-
-        pgs = self.client.placement.groups()
-
-        self.validate_pg_123(pgs[0])
-        assert pgs[0]._populated
-
-    def test_create_pg(self):
-        """
-        Tests that you can create a Placement Group.
-        """
-
-        with self.mock_post("/placement/groups/123") as m:
-            pg = self.client.placement.group_create(
-                "test",
-                "eu-west",
-                PlacementGroupType.anti_affinity_local,
-                PlacementGroupPolicy.strict,
-            )
-
-            assert m.call_url == "/placement/groups"
-
-            self.assertEqual(
-                m.call_data,
-                {
-                    "label": "test",
-                    "region": "eu-west",
-                    "placement_group_type": str(
-                        PlacementGroupType.anti_affinity_local
-                    ),
-                    "placement_group_policy": PlacementGroupPolicy.strict,
-                },
-            )
-
-            assert pg._populated
-            self.validate_pg_123(pg)
 
     def test_pg_assign(self):
         """
