+++ test/unit/objects/vpc_test.py
@@ -30,55 +30,6 @@ class VPCTest(ClientBaseCase):
         self.validate_vpc_123456(vpcs[0])
         self.assertEqual(vpcs[0]._populated, True)
 
-    def test_create_vpc(self):
-        """
-        Tests that you can create a VPC.
-        """
-
-        with self.mock_post("/vpcs/123456") as m:
-            vpc = self.client.vpcs.create("test-vpc", "us-southeast")
-
-            self.assertEqual(m.call_url, "/vpcs")
-
-            self.assertEqual(
-                m.call_data,
-                {
-                    "label": "test-vpc",
-                    "region": "us-southeast",
-                },
-            )
-
-            self.assertEqual(vpc._populated, True)
-            self.validate_vpc_123456(vpc)
-
-    def test_create_vpc_with_subnet(self):
-        """
-        Tests that you can create a VPC.
-        """
-
-        with self.mock_post("/vpcs/123456") as m:
-            vpc = self.client.vpcs.create(
-                "test-vpc",
-                "us-southeast",
-                subnets=[{"label": "test-subnet", "ipv4": "10.0.0.0/24"}],
-            )
-
-            self.assertEqual(m.call_url, "/vpcs")
-
-            self.assertEqual(
-                m.call_data,
-                {
-                    "label": "test-vpc",
-                    "region": "us-southeast",
-                    "subnets": [
-                        {"label": "test-subnet", "ipv4": "10.0.0.0/24"}
-                    ],
-                },
-            )
-
-            self.assertEqual(vpc._populated, True)
-            self.validate_vpc_123456(vpc)
-
     def test_get_subnet(self):
         """
         Tests that you can list VPCs.
@@ -124,32 +75,6 @@ class VPCTest(ClientBaseCase):
             )
 
             self.validate_vpc_subnet_789(subnet)
-
-    def test_list_ips(self):
-        """
-        Validates that all VPC IPs can be listed.
-        """
-
-        with self.mock_get("/vpcs/ips") as m:
-            result = self.client.vpcs.ips()
-
-        assert m.call_url == "/vpcs/ips"
-        assert len(result) == 1
-
-        ip = result[0]
-        assert ip.address == "10.0.0.2"
-        assert ip.address_range == None
-        assert ip.vpc_id == 123
-        assert ip.subnet_id == 456
-        assert ip.region == "us-mia"
-        assert ip.linode_id == 123
-        assert ip.config_id == 456
-        assert ip.interface_id == 789
-        assert ip.active
-        assert ip.nat_1_1 == "172.233.179.133"
-        assert ip.gateway == "10.0.0.1"
-        assert ip.prefix == 24
-        assert ip.subnet_mask == "255.255.255.0"
 
     def validate_vpc_123456(self, vpc: VPC):
         expected_dt = datetime.datetime.strptime(
